# backend/docs/swagger/swagger.json
{
    "basePath": "/api/v1",
    "definitions": {
...
        "ErrorResponse": {
            "properties": {
                "code": {
                    "example": "TEAM_NOT_FOUND",
                    "type": "string"
                },
                "error": {
                    "example": "Team not found",
                    "type": "string"
                },
                "message": {
                    "example": "No team exists with id team-unknown",
                    "type": "string"
                }
            },
            "type": "object"
        },
...
        "TeamSubmissionStatusResponse": {
            "properties": {
                "allSubmitted": {
                    "example": true,
                    "type": "boolean"
                },
                "assessmentPeriod": {
                    "example": "2025 - 1st Half",
                    "type": "string"
                },
                "postWorkshopExists": {
                    "example": false,
                    "type": "boolean"
                },
                "submittedMembers": {
                    "example": 5,
                    "type": "integer"
                },
                "teamId": {
                    "example": "team-phoenix",
                    "type": "string"
                },
                "totalMembers": {
                    "example": 4,
                    "type": "integer"
                }
            },
            "type": "object"
        },
...
    },
    "info": {
        "contact": {},
        "description": "Team360 Health Check Application REST API. Assesses team health across 11 dimensions with hierarchical org support.",
        "title": "Team360 API",
        "version": "1.0"
    },
    "paths": {
...
        "/assessment-periods": {
            "get": {
                "description": "Returns the distinct assessment period values that appear on at least one submitted health check session, for populating period-selector dropdowns in the frontend.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "periods",
                        "schema": {
                            "additionalProperties": {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch assessment periods",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List distinct assessment periods",
                "tags": [
                    "Health Checks"
                ]
            }
        },
...
        "/teams/{teamId}/submission-status": {
            "get": {
                "description": "Returns the submission progress for the given team and assessment period (both required query/path values): total team members, how many have submitted, whether every member has submitted, and whether a post-workshop survey exists for that period. Fails with a 400 if either the team ID or assessment period is missing.",
                "parameters": [
                    {
                        "description": "Team ID",
                        "in": "path",
                        "name": "teamId",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "Assessment period (e.g. '2024 - 1st Half')",
                        "in": "query",
                        "name": "assessmentPeriod",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/TeamSubmissionStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Missing teamId or assessmentPeriod",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get submission status",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Get team's submission status",
                "tags": [
                    "Health Checks"
                ]
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT token.",
            "in": "header",
            "name": "Authorization",
            "type": "apiKey"
        }
    },
    "swagger": "2.0"
}